requirements.txt
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==3.2
djangorestframework==3.13.1
Faker==13.14.0
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.4
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.18.1
watchfiles==0.15.0
websockets==10.3
zipp==3.8.0
